# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/create-multiple-wallet": {
            "post": {
                "description": "创建具有指定地址的多重签名钱包。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "钱包"
                ],
                "parameters": [
                    {
                        "description": "用于创建多重签名钱包的请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateMultipleWallet"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.CreateMultipleWalletResponse"
                        }
                    }
                }
            }
        },
        "/deleteMember": {
            "delete": {
                "description": "删除指定钱包地址的成员",
                "produces": [
                    "application/json"
                ],
                "summary": "删除成员",
                "parameters": [
                    {
                        "type": "string",
                        "description": "钱包地址",
                        "name": "wallet_address",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "成员地址",
                        "name": "address",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "request.CreateMultipleWallet": {
            "type": "object",
            "required": [
                "address",
                "is_weighted",
                "threshold"
            ],
            "properties": {
                "address": {
                    "description": "地址",
                    "type": "string"
                },
                "is_weighted": {
                    "description": "是否为权重值钱包",
                    "type": "integer"
                },
                "threshold": {
                    "description": "多签门限值",
                    "type": "integer"
                }
            }
        },
        "response.CreateMultipleWalletResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                }
            }
        }
    }
}